divide todolist view logic
